In symbol selector, omit decimals when displaying the transparency %
git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@13324 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/gui/symbology-ng/qgssymbolv2selectordialog.cpp
@@ -260,5 +260,5 @@ void QgsSymbolV2SelectorDialog::on_mTransparencySlider_valueChanged( int value )
 void QgsSymbolV2SelectorDialog::displayTransparency( double alpha )
 {
   double transparencyPercent = ( 1 - alpha ) * 100;
-  mTransparencyLabel->setText( tr( "Transparency: %1%" ).arg( transparencyPercent ) );
+  mTransparencyLabel->setText( tr( "Transparency: %1%" ).arg(( int ) transparencyPercent ) );
 }
